service-python/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
tensorflow
pillow
numpy
pandas
scikit-learn
xgboost
joblib
